refactor: TagRating을 Tag에 병합 TagRating 삭제
